moving version to its own component
