+++ source/commands/Upgrade.tsx
@@ -8,34 +8,37 @@ import {ExecaError, execa} from 'execa';
 import constants from '../constants.js';
 
 export default () => {
-	const {latest} = useCliContext();
+	const {version} = useCliContext();
 	const running = useBoolean(true);
 	const [error, setError] = useState<string | undefined>();
 
 	useEffect(() => {
-		upgrade();
-	}, []);
+		if (version.latest) {
+			upgrade();
+		}
+	}, [version.latest]);
 
 	const upgrade = async () => {
-		try {
-			await execa(
-				'npm',
-				['upgrade', '-g', `${constants.package_name}@${latest}`],
-				{
-					shell: true,
-				},
-			);
-		} catch (e) {
-			setError((e as ExecaError).message);
-		} finally {
-			running.setFalse();
-		}
+		if (version)
+			try {
+				await execa(
+					'npm',
+					['upgrade', '-g', `${constants.package_name}@${version.latest}`],
+					{
+						shell: true,
+					},
+				);
+			} catch (e) {
+				setError((e as ExecaError).message);
+			} finally {
+				running.setFalse();
+			}
 	};
 
 	if (running.value) {
-		return <Spinner text={`Upgrading Courier CLI ${latest}`} />;
+		return <Spinner text={`Upgrading Courier CLI ${version.latest}`} />;
 	} else {
-		const text = `Courier CLI upgraded to ${latest}`;
+		const text = `Courier CLI upgraded to ${version.latest}`;
 		return (
 			<Box width={error ? undefined : text.length + 8}>
 				{error?.length ? (
